Moved date formatter to helper file
